remodelled actiosn structure
