+++ client/src/reducers/loaderReducer.js
@@ -9,9 +9,18 @@ const toastReducer = (state = initialState, action) => {
         case actionTypes.LOGIN_USER_STARTED:
             return{ ...state, loading: true}
         case actionTypes.LOGIN_USER_FAILED:
-            return{ ...state, loading: false}
         case actionTypes.LOGIN_USER_SUCESS: 
-        return{ ...state, loading: false}
+            return{ ...state, loading: false}
+        case actionTypes.REGISTERATION_STARTED:
+            return{ ...state, loading: true}
+        case actionTypes.REGISTRATION_FAILED:
+        case actionTypes.REGISTRATION_SUCCESS: 
+            return{ ...state, loading: false}
+        case actionTypes.LOGOUT_USER_STARTED:
+            return{ ...state, loading: true}
+        case actionTypes.LOGOUT_USER_SUCESS:
+        case actionTypes.LOGIN_USER_FAILED:
+            return{ ...state, loading: false}
         default:
             return state
     }
