+++ client/src/reducers/toastReducer.js
@@ -8,8 +8,13 @@ const toastReducer = (state = initialState, action) => {
     switch (action.type) {
         case actionTypes.LOGIN_USER_SUCESS:
             return{ ...state, toastMessage: "User Successfully Logged In"}
-        case actionTypes.REGISTRATION_FAILED:
+        case actionTypes.LOGIN_USER_FAILED:
+            return{ ...state, toastMessage: action.error}
+        case actionTypes.REGISTRATION_SUCCESS:
             return{ ...state, toastMessage: action.payload}
+        case actionTypes.LOGOUT_USER_SUCESS:
+            return{ ...state, toastMessage: action.payload}
+        
         case actionTypes.RESET_TOAST: 
             return{ ...state, toastMessage: ""}
         default:
